vfs: Move vfs_syscalls into own file
Also implement FileHandle struct which right now only has a reference to a vnode.
Add method to task so we can add a FileHandle struct to the task.

Remove error file in fs module, move error enum into mod.rs

Impl from fs::error for syscallerror, so we can use fs::error in filesystem and not have to use syscallerror enum

Use match instead of if in lookup for tmpfsnode